pyproject.toml
[tool.poetry]
name = "tera-erp"
version = "0.1.0"
description = ""
authors = ["dennis marcus goh"]
license = "LGPLv3"
readme = "README.md"
packages = [
    { include = "app", from = "." }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.124.0"
sqlalchemy = "^2.0.44"
uvicorn = "^0.38.0"
asyncpg = "^0.31.0"
pydantic = {extras = ["email"], version = "^2.12.5"}
pydantic-settings = "^2.12.0"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
email-validator = "^2.1.0"
pyyaml = "^6.0.3"
alembic = "^1.17.2"
reportlab = "^4.4.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
